requirements.txt
numpy >= 2; python_version>='3.9'
numpy < 2; python_version<'3.9'
#PIL

#FontTools-TTX
http://sourceforge.net/projects/fonttools/files/2.3/fonttools-2.3.tar.gz/download
TTFQuery

PyOpenGL
PyOpenGL-accelerate

pydispatcher
PyVRML97
PyVRML97-accelerate

simpleparse
